+++ Modules/ProfileFever/ProfileFever.psm1
@@ -47,6 +47,7 @@ $Script:PromptAlias    = $false
 $Script:PromptTimeSpan = $false
 $Script:PromptGit      = $false
 $Script:PromptDefault  = Get-Command -Name 'prompt' | Select-Object -ExpandProperty 'Definition'
+$Script:PromptTitle    = $null
 
 # Enumerate the prompt color based on the operating system
 if ([System.Environment]::OSVersion.Platform -eq 'Win32NT')